+++ mayan/apps/file_metadata/managers.py
@@ -18,9 +18,7 @@ class ModelManagerDocumentTypeDriverConfigurationValid(models.Manager):
     def get_queryset(self):
         queryset = super().get_queryset()
 
-        return queryset.filter(
-            stored_driver__exists=True
-        )
+        return queryset.filter(stored_driver__exists=True)
 
 
 class ModelManagerStoredDriverValid(models.Manager):