+++ mayan/apps/file_metadata/tests/test_property_helpers.py
@@ -1,14 +1,14 @@
 from mayan.apps.common.tests.mixins import PropertyHelperTestMixin
 from mayan.apps.testing.tests.base import BaseTestCase
 
-from .mixins import DocumentFileMetadataTestMixin
+from .mixins.document_file_mixins import DocumentFileMetadataTestMixin
 
 
 class DocumentPropertyHelperTestCase(
     DocumentFileMetadataTestMixin, PropertyHelperTestMixin, BaseTestCase
 ):
-    _test_document_file_metadata_entry_create_auto = True
     _test_document_file_metadata_driver_create_auto = True
+    _test_document_file_metadata_entry_create_auto = True
 
     def test_basic(self):
         self._clear_events()
@@ -16,8 +16,8 @@ class DocumentPropertyHelperTestCase(
         self.assertEqual(
             getattr(
                 self._test_document.file_metadata_value_of,
-                self._test_document_file_metadata_path
-            ), self._test_document_file_metadata.value
+                self._test_document_file_metadata_entry_path
+            ), self._test_document_file_metadata_entry.value
         )
 
         events = self._get_test_events()
@@ -31,7 +31,7 @@ class DocumentPropertyHelperTestCase(
         self.assertEqual(
             getattr(
                 self._test_document.file_metadata_value_of,
-                self._test_document_file_metadata_path
+                self._test_document_file_metadata_entry_path
             ), None
         )
 
@@ -39,14 +39,14 @@ class DocumentPropertyHelperTestCase(
         self.assertEqual(events.count(), 0)
 
     def test_no_file_metadata(self):
-        self._test_document_file_metadata.delete()
+        self._test_document_file_metadata_entry.delete()
 
         self._clear_events()
 
         self.assertEqual(
             getattr(
                 self._test_document.file_metadata_value_of,
-                self._test_document_file_metadata_path
+                self._test_document_file_metadata_entry_path
             ), None
         )
 
@@ -54,7 +54,7 @@ class DocumentPropertyHelperTestCase(
         self.assertEqual(events.count(), 0)
 
     def test_wrong_file_metadata_name(self):
-        self._test_document_file_metadata.delete()
+        self._test_document_file_metadata_entry.delete()
 
         self._clear_events()
 
@@ -72,8 +72,8 @@ class DocumentPropertyHelperTestCase(
 class DocumentFilePropertyHelperTestCase(
     DocumentFileMetadataTestMixin, PropertyHelperTestMixin, BaseTestCase
 ):
-    _test_document_file_metadata_entry_create_auto = True
     _test_document_file_metadata_driver_create_auto = True
+    _test_document_file_metadata_entry_create_auto = True
 
     def test_basic(self):
         self._clear_events()
@@ -81,22 +81,22 @@ class DocumentFilePropertyHelperTestCase(
         self.assertEqual(
             getattr(
                 self._test_document_file.file_metadata_value_of,
-                self._test_document_file_metadata_path
-            ), self._test_document_file_metadata.value
+                self._test_document_file_metadata_entry_path
+            ), self._test_document_file_metadata_entry.value
         )
 
         events = self._get_test_events()
         self.assertEqual(events.count(), 0)
 
     def test_no_file_metadata(self):
-        self._test_document_file_metadata.delete()
+        self._test_document_file_metadata_entry.delete()
 
         self._clear_events()
 
         self.assertEqual(
             getattr(
                 self._test_document_file.file_metadata_value_of,
-                self._test_document_file_metadata_path
+                self._test_document_file_metadata_entry_path
             ), None
         )
 
@@ -104,7 +104,7 @@ class DocumentFilePropertyHelperTestCase(
         self.assertEqual(events.count(), 0)
 
     def test_wrong_file_metadata_name(self):
-        self._test_document_file_metadata.delete()
+        self._test_document_file_metadata_entry.delete()
 
         self._clear_events()